Changing Direction
This project will no longer be a SWAT simulator, but rather a general police simulator. Players will be able to respond to calls from a Ready or Not style map. Each call will be procedurally generated but each call will be based on a template.
- Player Movement
The player can crouch, walk, and lean. Movement based on old project
- Doors
The player can peek and swing open doors
- Layout Generator
A script can choose random locations on a map and then place random objects at those locations
- Weapon Data
- Started Development of Weapons
- First weapon 3d model